requirements.txt
paho-mqtt
wyzesense
